+++ phpunit.xml.dist
@@ -1,13 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<phpunit colors="true" bootstrap="tests/bootstrap.php">
+<phpunit colors="true" bootstrap="vendor/behat/mink/driver-testsuite/bootstrap.php">
     <testsuites>
         <testsuite name="Driver test suite">
-            <directory>./tests/Behat/Mink/Driver</directory>
+            <directory>tests</directory>
+            <directory>vendor/behat/mink/driver-testsuite/tests</directory>
         </testsuite>
     </testsuites>
 
     <php>
+        <var name="driver_config_factory" value="Behat\Mink\Tests\Driver\Selenium2Config::getInstance" />
+
         <!--server name="WEB_FIXTURES_HOST" value="http://test.mink.dev" /-->
         <!--server name="WEB_FIXTURES_BROWSER" value="firefox" /-->
