--- tests/bootstrap.php
@@ -1,8 +1,0 @@
-<?php
-
-if (!is_dir($vendor = __DIR__.'/../vendor')) {
-    die('Install dependencies first');
-}
-
-require($vendor.'/autoload.php');
-require($vendor.'/behat/mink/tests/Behat/Mink/Driver/CssDriverTest.php');
